Pathfinder.py
messing with the tkinter module was wasting time, was able to create the same functionality in half the lines of code. Will revisit to add a way to update the user as the pathping is generated, also the main goal will be to print out pathping to a map, which will be hard.